Refactor etad module to sit under ancillary module
